Righten a node.java file
fix..

+++ src/dsaproject/Node.java
@@ -25,7 +25,7 @@ public class Node {
     
     }
 
-    Node(String title, Long bn, String authorFName, String authorSName) {
+    Node(String title, int bn, String authorFName, String authorSName) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
